move ResultType object to Constants.kt file and move Constants.kt file to util package
