--- app/src/main/java/com/sadegh/calculator/homeScreen/ResultType.kt
@@ -1,10 +1,0 @@
-package com.sadegh.calculator.homeScreen
-
-import androidx.compose.ui.res.painterResource
-import com.example.calculator.R
-
-object ResultType {
-
-    const val undefined = "can not divide by zero"
-
-}
